Move LH5 tools to dedicated module
